+++ osu.Game.Tests/Visual/SongSelectV2/TestSceneLeaderboardScore.cs
@@ -20,7 +20,7 @@ using osu.Game.Rulesets.Osu;
 using osu.Game.Rulesets.Osu.Mods;
 using osu.Game.Rulesets.Scoring;
 using osu.Game.Scoring;
-using osu.Game.Screens.SelectV2.Leaderboards;
+using osu.Game.Screens.SelectV2;
 using osu.Game.Tests.Resources;
 using osu.Game.Users;
 using osuTK;
@@ -60,7 +60,7 @@ namespace osu.Game.Tests.Visual.SongSelectV2
 
                 foreach (var scoreInfo in getTestScores())
                 {
-                    fillFlow.Add(new LeaderboardScoreV2(scoreInfo)
+                    fillFlow.Add(new BeatmapLeaderboardScore(scoreInfo)
                     {
                         Rank = scoreInfo.Position,
                         IsPersonalBest = scoreInfo.User.Id == 2,
@@ -93,7 +93,7 @@ namespace osu.Game.Tests.Visual.SongSelectV2
 
                 foreach (var scoreInfo in getTestScores())
                 {
-                    fillFlow.Add(new LeaderboardScoreV2(scoreInfo)
+                    fillFlow.Add(new BeatmapLeaderboardScore(scoreInfo)
                     {
                         Rank = scoreInfo.Position,
                         IsPersonalBest = scoreInfo.User.Id == 2,
@@ -108,7 +108,7 @@ namespace osu.Game.Tests.Visual.SongSelectV2
         [Test]
         public void TestUseTheseModsDoesNotCopySystemMods()
         {
-            LeaderboardScoreV2 score = null!;
+            BeatmapLeaderboardScore score = null!;
 
             AddStep("create content", () =>
             {
@@ -146,7 +146,7 @@ namespace osu.Game.Tests.Visual.SongSelectV2
                     Date = DateTimeOffset.Now.AddYears(-2),
                 };
 
-                fillFlow.Add(score = new LeaderboardScoreV2(scoreInfo)
+                fillFlow.Add(score = new BeatmapLeaderboardScore(scoreInfo)
                 {
                     Rank = scoreInfo.Position,
                     Shear = Vector2.Zero,
