+++ osu.Game/Screens/OnlinePlay/DailyChallenge/DailyChallengeLeaderboard.cs
@@ -17,7 +17,7 @@ using osu.Game.Online.Rooms;
 using osu.Game.Rulesets;
 using osu.Game.Rulesets.Mods;
 using osu.Game.Scoring;
-using osu.Game.Screens.SelectV2.Leaderboards;
+using osu.Game.Screens.SelectV2;
 using osuTK;
 
 namespace osu.Game.Screens.OnlinePlay.DailyChallenge
@@ -40,7 +40,7 @@ namespace osu.Game.Screens.OnlinePlay.DailyChallenge
         private readonly Room room;
         private readonly PlaylistItem playlistItem;
 
-        private FillFlowContainer<LeaderboardScoreV2> scoreFlow = null!;
+        private FillFlowContainer<BeatmapLeaderboardScore> scoreFlow = null!;
         private Container userBestContainer = null!;
         private SectionHeader userBestHeader = null!;
         private LoadingLayer loadingLayer = null!;
@@ -91,7 +91,7 @@ namespace osu.Game.Screens.OnlinePlay.DailyChallenge
                                 new OsuScrollContainer
                                 {
                                     RelativeSizeAxes = Axes.Both,
-                                    Child = scoreFlow = new FillFlowContainer<LeaderboardScoreV2>
+                                    Child = scoreFlow = new FillFlowContainer<BeatmapLeaderboardScore>
                                     {
                                         RelativeSizeAxes = Axes.X,
                                         AutoSizeAxes = Axes.Y,
@@ -158,7 +158,7 @@ namespace osu.Game.Screens.OnlinePlay.DailyChallenge
                 }
                 else
                 {
-                    LoadComponentsAsync(best.Select((s, index) => new LeaderboardScoreV2(s, sheared: false)
+                    LoadComponentsAsync(best.Select((s, index) => new BeatmapLeaderboardScore(s, sheared: false)
                     {
                         Rank = index + 1,
                         IsPersonalBest = s.UserID == api.LocalUser.Value.Id,
@@ -178,7 +178,7 @@ namespace osu.Game.Screens.OnlinePlay.DailyChallenge
 
                 if (userBest != null)
                 {
-                    userBestContainer.Add(new LeaderboardScoreV2(userBest, sheared: false)
+                    userBestContainer.Add(new BeatmapLeaderboardScore(userBest, sheared: false)
                     {
                         Rank = userBest.Position,
                         IsPersonalBest = true,
